requirements.txt
beautifulsoup4==4.6.0
selenium==3.14.0
pymongo==3.9.0
termcolor==1.1.0
mysql-connector-python==1.16.0
